base-app/src/backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
pillow==11.0.0